FindMessage is used when a reply is expected for a sent message.
As an example, IPC_REQUEST from the client expects the specific reply
of IPC_REQUEST_REPLY. It is possible that the server sneaks in other
messages, such as a stream of IPC_ADVISE data. So FindMessage loops
until the desired message is found, executing any other unexpected
messages that happen to be read first.